refactor to loop

+++ i2c.go
@@ -2,7 +2,7 @@ package firmata
 
 // I2cReply represents the response from an I2cReply message
 type I2cReply struct {
-	Address  int
+	Address  byte
 	Register int
 	Data     []byte
 }
